Fix: Beat 计算时小数丢失

+++ Cyan-Stars/Assets/Scripts/Chart/Beat.cs
@@ -85,7 +85,7 @@ namespace CyanStars.Chart
                 return IntegerPart;
             }
 
-            return IntegerPart + Numerator / Denominator;
+            return IntegerPart + Numerator / (double)Denominator;
         }
